pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "github-readme-stats-cache"
description = 'Caches GitHub Readme Stats images served by a Vercel app.'
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
    { name = "Tuomo Kriikkula", email = "tuokri@tuta.io" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "aiohttp[speedups]>=3.8.3",
    "certifi>=2022.12.7",
    "python-dotenv>=0.21.0",
    "redis>=4.4.0",
    "sanic>=22.9.1",
    "uvloop>=0.17.0; platform_system != 'Windows'",
]
dynamic = ["version"]

[project.urls]
Documentation = "https://github.com/unknown/github-readme-stats-cache#readme"
Issues = "https://github.com/unknown/github-readme-stats-cache/issues"
Source = "https://github.com/unknown/github-readme-stats-cache"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "_version.py"

#[tool.hatch.envs.default]
#dependencies = [
#  "pytest",
#  "pytest-cov",
#]
#[tool.hatch.envs.default.scripts]
#cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=github_readme_stats_cache --cov=tests"
#no-cov = "cov --no-cov"

[[tool.hatch.envs.test.matrix]]
python = ["38", "39", "310", "311"]

#[tool.coverage.run]
#branch = true
#parallel = true
#omit = [
#  "github_readme_stats_cache/__about__.py",
#]

#[tool.coverage.report]
#exclude_lines = [
#  "no cov",
#  "if __name__ == .__main__.:",
#  "if TYPE_CHECKING:",
#]
